+++ tests/gdi/test_protected_graphics.cpp
@@ -54,7 +54,7 @@ namespace
             }
 
             void rdp_input_scancode(KbdFlags /*flags*/, Scancode /*scancode*/, uint32_t /*event_time*/, Keymap const& /*keymap*/) override {}
-            void rdp_input_mouse(int /*device_flags*/, int /*x*/, int /*y*/) override {}
+            void rdp_input_mouse(uint16_t /*device_flags*/, uint16_t /*x*/, uint16_t /*y*/) override {}
             void rdp_input_synchronize(KeyLocks /*locks*/) override {}
 
             // Client Notify module that gdi is up and running
